+++ src/sql/workbench/contrib/notebook/browser/calloutDialog/media/calloutDialog.css
